requirements.txt
torch
torchvision
numpy
pandas
matplotlib
opencv-python
scikit-image
scikit-learn
tqdm
pyyaml
Pillow
jupyter